Always require authentication to accept invitations, fixes #731
The old invitation form was giving the user options to login or register in
case he was not logged or registered yet when trying to accept and invivation
to join a space. Now we always require authentication for the user to accept
an invivation.
It's simpler and the user can simply login or register using the standard
views. It works better when the application has different methods to login
or doesn't allow registration.