Fix emails being sent with the wrong mail address, fix #713
All emails have to be sent with the email configured in the Site. Spams
and feedbacks were being sent with the email of the user that created the
spam/feedback. This fails for some SMTP servers and is wrong.

cherry-pick 1696f298e8a025df552ab63e55ee3a76abd4651f