Better rescue from Station::NotAuthorized as station is updated
Now station will always throw Station::NotAuthorized so we have to treat it
and render the pages we need to.
Spaces and routes inside spaces will redirect to the new join request page
if the user is trying to *see* a page. If he's trying to do something else
(e.g. update a post he can't updated), he will still get a 403.