Remove lots of vendored assets that are not used anymore
Most of them are not referenced anywhere in the code anymore.
A few of them are still referenced in old files. But if they are needed
in the future we will have to find better alternatives or at least
update their versions (most of them were more than 2 years old).