Yield more than just :headers in layouts, put assets in the right place
Now there's a specific place for javascripts, stylesheets and general headers.
Previously when using the layout "no_sidebar" a css was being included at
the bottom of the document, causing the page to be shown initially with a
space for the sidebar and making the other components to be placed incorrectly.