Fix permission error fetch recordings on /homes/recordings
The view was submitting a POST to an .html instead of a .json and
the controllers was trying to redirect to a route that the user has
no access to.
Also improved the "effect" applied to the div with the recordings. Now
the effect is only shown if there is at least one recording.

fixes #908