Add recording options to user's rooms, refs #804
First version where the user can set his room to record and set the title and
description for the recording. Still only for user rooms.

Added also a new js component "edit in place", that allows inputs to be easily
submitted via ajax to update a model. This can possibly be used in the future
to make other forms more responsive and easier to be edited.