Remove old and outdated locale files for some languages
Removing 'fr', 'de' and 'es', languages that were available initially in vcc
but were never updated and are now way behind the other languages and are only
a burden in the development.

Should we get these languages properly updated in the future then they will be
added back to the application.