Fix empty password inputs when editing models, refs #949
It was actually the expected behavior to have an empty value in a password
input. But in these inputs we actually want to show the values, so we have
to explicitly set it when creating the input.
This will be valid for inputs of the type 'showable_password'. For standard
password inputs the value won't be displayed (and they shouldn't).