Move routes nested at /home to /my
Mostly a renaming of HomesController to MyController. The actions that are there
make more sense under a route such as /my/home and /my/activity. They are routes
available to the current user only, but not necessarily related to his home.
More actions should be added in the future in this controller.

Note that the route helpers for these actions changed!