Use `description` in recordings to describe them, refs #986
Added migrations to remove the metadata from rooms and to move the content
from the metadata in the recordings to the `description` attribute.
Also added routes and views for the user to edit the description of a
recording when it is made available.

The permission control is ok but not 100%. A few TODOs were added in
Abilities for things that should be reviewed.