Started the implementation of shibboleth in mweb2, refs #376
Very first changes, migrating a bit of what was done in the master branch.
Not only copying the code, but improving it as much as possible.
Still a lot to be done, but already included tests for most of what was done
so far.

Had to add the migrations as done in master (using the same timestamps) so
the migration from the current master to branch-v2 will work (refs #987).