Fixed issue where all spaces are loaded together
Inside the index view page of spaces, all spaces are loaded together, this generate a overload who is transmitted to the user in form of slowness. Fixed this by adding pagination to this page, and was created a way to user find any space by searching by name or permalink. Some styles were changed to become padronized with the rest of site pages.