Fix undefined method 'default' for BigbluebuttonServer
Was hapenning in development only, because rails unloads initializers and
doesn't reaload them again. Fixed adding the initializer inside a 'to_prepare'
block.
Note that it won't work with 'ActiveSupport.on_load(:after_initialize)' as
done in some other initializers.

More at: http://stackoverflow.com/questions/5852580/why-do-includes-in-rails-engine-initializers-malfunction-when-cache-classes-fahttp://stackoverflow.com/questions/5852580/why-do-includes-in-rails-engine-initializers-malfunction-when-cache-classes-fahttp://stackoverflow.com/questions/5852580/why-do-includes-in-rails-engine-initializers-malfunction-when-cache-classes-fa