Add option to require admins to accept new registrations
It allows admins to set a flag in the site that will require admins to accept
new registrations. The users are allowed to register new accounts, but they will
only be allowed to login once they are accepted.
By default this flag is off, so the site behaves as before: registration is open.

This was implemented in mconf-rnp/mconf-web, and this implmementation is based
on that, with a few more features.

refs #953