Update MwebEvents and require its assets properly
Now all partials rendered by the layouts must include `main_app.` before any
route path helper that is called. Since these layouts are used when rendering
the views on MwebEvents, we have to add `main_app.` to tell rails that the route
helpers are from the main application and not from the gem.
Should try to find a better solution in the future.

refs #1115