Fixed the problem when a user/space is created with the same identifier of another
When a user or a space are created the system try to create his room using the username/permalink like param. This can result in a error, because the room can't be created with a param equal to another. To resolve this issue a validate function was created in user/space model to verify if doesn't exist another model with the same identifier. If another identifier is found the model aren't created.

Fixed #1071