Created a new way to allow users to select the default layout of a room
Before this modification was impossible to change the layout of a room, all rooms were presented with the same default layout. Now the idea is to let owners/admins choose from a pre selected list of layouts. For do this was inserted in the configuration pages  of the room (space and user rooms) a new field who owners/admins can select the default layout to be used.

Fix #960