Fix wrong language in space invitation emails
The same that was done in the master branch by e3ebbc466756b17.

Was sending the message in the wrong language in some cases, and in some cases
using more than one language in the same email.
This was in the email sent to the admins of a space confirming that a user
accepted or rejected an invitation to join the space.

refs #984