Make the event module really optional
It shouldn't be installed by default, only if the user wants to.
The application should run properly either with or without the module.

Now it is optional in the Gemfile and all references to MwebEvents and its
classes are protected to run only if the module was loaded. The initializer
now sets flags to tell the application if the module is loaded/enabled.

refs #1115