refs #1076
refs #1077

Removed the ldap authentication option from login view.
Made sure that ldap authentication is the first to be executed if it is enabled on the site.
Made sure that if ldap is enabled and the authentication process fails, it executes the standard authentication.

refs #1055
Added more info about the ldap authentication process to the log.