Refactor private_messages/new to use select2
It also pre populates the input with the user selected, it any. With this we
can click a link to send a message to an specific user and the select2 input
will be automatically set with this target user.

Had to improve a bit the route users/select to search users by id.

refs #275