Improve how we check if the event module is loaded/enabled
Now the basic functions (routes, abilities, associations in models) are always
set if the module is *loaded*. If the event is *disabled*, the routes will
exist, but return a 404, and the views will filter what is shown.

Now we can enable and disable the module without having to restart the
application.
Also, the previous solution was causing errors in production, so had to be
improved.

Complements 8b7d4ab9c0 and 0627dce8c60

refs #1115