Fix error messages not being shown in the registration form
For some reason the `resource` gets in the view as nil when the page is loaded
via registrations#new. Had fixed this in be47d4f83a665, but then the errors
when a registration failed where not being shown anymore. Happened because
the form was always creating a new user always and not keeping the one
submitted to registrations#create.