Add new attribute "principal name" to shibboleth fields
It's an attribute used by shibboleth to store the user's unique identifier.
Can be used in the application to get an identifier for the user or the user's
institution. It is required, otherwise the login will fail.

Also removed the default attribute names that were used in case they were
not configured in the website. Now they will always have to be configured.

refs #1230