First changes for the new logic when joining via mobile
There will be a single button to join a conference. If the user is in a mobile
device, the mobile client will be automatically triggered. The old modal window
to join via mobile was removed.

Made a new "mobile" layout, currently only used in the page shown to join a
conference from a mobile client, but that can be used for more in the future.