Improvements after PR #19
The most important was the removal of API calls to the webconf server from
Abilities. Now these calls are made by the controller and Abilities will only
check for information already fetched into the models.
Updated bigbluebutton_rails to have this change in effect.

The rest was a bit of refactoring in the changes from the PR.

refs #1148