New js component DateTimeInput to fix the invitation form
The invitation form now uses multiple inputs (<select>s) to select a time
and a datetimepicker to select a date (the same way is done for events in
mweb_events).
These inputs are grouped together in a new js class DateTimeInput.
Not finished yet, but the js part of the form is almost finished.