Refactor the feedback email and the tests for ApplicationMailer
New feedback email and form, using haml, simple form and everything else we
use in Mconf-Web v2.

Tests for mailer methods should only test that the email returned has the
correct properties and body set. Tests that the mail is sent/enqueued should
be made elsewhere. For the digest_email, for instance, the test that it is
enqueued is at spec/lib/mconf/digest_email_spec, since this is the class that
actually calls the "deliver".

refs #371