Refactor failing migration after rails 4
Was failing because it tried to use Space#save, that tried to also save its
associated BigbluebuttonRoom, that failed for having the attribute #record
that's invalid in rails 4.
Had to use fake models instead. Couldn't use raw SQL because we also need to
generate the images according to CarrierWave, and it would not be ideal to
do it all manually.

refs #987