Remove broken page view statistics
Page views for spaces used to be taken from google analytics.
It was not being used anymore for a long time and would require some work
to get it back working. The feature is actually not necessary right now,
so it was removed.
Might be added again in the future if necessary.

refs #977