Tests for confirmation email, with changes in the test suite for it
A continuation of 5125f60.
Had to make some changes in the test suite to be able to send and test emails
with devise and resque.
Disabled use_transactional_fixtures and started using DatabaseCleaner to
clean up the test database. Most of the times it will still use transactions,
but it can be changed to truncation if we need it.

refs #1322