Fix errors on error pages' css, no more animations and cleaner style
Updated compass, the version we were using was old. There is an even newer
version that supports animations, but it is not compatible with the current
bootstrap-sass. So no animations for now, but it can be added again in the
future when compass is updated.
Made the design cleaner too to match the new site design better.

refs #1087