Fix permissions for users in conferences
The order of the methods to load the room, fetch information and check
permissions was not exactly right. It was also loading the room in multiple
places and fetching the information into one variable, but checking the
permissions using another.

Added a few more tests to try to make sure that things run in the order they
should and using the right variables.

refs #1376