Fix handling of errors, complements #46
The error handling class should not be inside a Mailer, because it ends up
never being called. Rails does something funny with mailer classes, so the
call we had to mailer.error_handler was not triggering the method as it
should.
Ended up moving the method to a different class and now everything works
fine and got even easier to test.

refs #1365