Better validators for username and permalink uniqueness
It was working for most of the cases, except when we tried to create a user
or space with a username/permalink that would *not* conflict with other user
or space, but that would generate a room that would conflict with another
room.

Created validator classes to do proper validations over usernames, permalinks
and room params.

refs #609