Fixing more possible exceptions in workers and renaming some of them
Objects used by the workers are not always 100% complete when the worker
is running, so we have to check this cases and just print log messages instead
of breaking the worker.
One example is when a user invites someone to join a space but removes the
invitation before the worker runs. When the worker runs, the invitation will
have no associated join request.

Also renamed the workers to make it more clear what they do. Workers with
"Sender" are the ones that actually send notifications. The others are the
ones that go through a list of objects and schedule a sender for each.