Improve how join requests are shown, accepted and declined
The view that shows a join request is now more complete and treats all
possible cases in more detail (invitation, request, who requested, etc).

Several changes in the controller to use the actions 'accept' and 'decline'
instead of 'update'. These are more specific actions and with them the code
is cleaner and we can restrict more the attributes the users need access to.

Also reviewed the recent activity items being created (partly 42e7173bce150,
that's not integrated yet).

refs #1310