Move user approval notifications from the model to workers
Instead of sending emails from after_create and after_update callbacks in
User, now there are a few resque workers that detect when approval emails
need to be sent and send them.

This fix the fact that sometimes the worker called to send an email was
running even before the user was persisted to the database, generating an
error in the worker.

refs #1403