Fix display of accepted join requests after PR #51
Accepted or declined join requests now have the space set as "trackable"
instead of "owner", so they were not being displayed in the list of
recent activities of a user.
Made new methods to get only the public activities: all the activities that
were already being returned, plus the "user joined space". It rejects the
"user declined" activities.

refs #1395