More flexible control over img sizes when cropping
Before, was assuming the images were always in the resolution 350x350.
Now the cropping can show the image in any size and the coordinates will
be relative to this size. We still show and crop images to a max of 350x350,
but the control over this size is more flexible now.

The original image is now stored cropped, so if we need to generate new
version in the future we can. Currently we store the entire image (uncropped)
and we don't store the cropping coordinates so there's no way to generate
new versions.
Also added larger image formats, not used right now but we will need them in
the future.


refs #1051