Make sure confirmation routes show a 404 when registrations are disabled
When the flag site_registration_enabled was false, it would throw a 500
in production when the user accessed any of the confirmation routes
(e.g. /users/confirmation/new).
Also, the action confirmations#show was not redirecting the user to his
home if the user is already confirmed.
Fixed both errors.

refs #1413