Fix failing commands to stop resque workers
Now we trigger multiple calls to resque:work instead of a single call to
resque:workers. With this we can manage the pids and stop the processes
correctly and monitor the processes individually.
Made the calls to resque:work run in background, so the process that
calls the script start_resque_workers.sh is free to finish.

Also improved the format of the logs written by resque.

refs #1424
refs #1391