Explicitly restrict actions over disabled resource in abilities
Disabled resources and their nested resources were already being restricted
on controllers and views by the default scopes set on User and Space. But
there wasn't much in the abilities to specify that users should not be
able to access anything on them.
Included this part that was missing to explicitly restrict access over
disabled resources on the model level.

refs #1173