Rename the helper that renders "time ago"s and make it easier to use
The helper DatesHelper#time_to_be_shown is now DatesHelper#formatted_time_ago
and it does more than before. Now it can be called with a single date and
it will automatically set the text and title of the <span> created.

Also made the title easier to read for users (to_formatted_s(:long)).

refs #1307