Add es-419 (latin america) and a way to set the visible languages
There's a new option on Site to set what languages should be visible.
With this we can add more languages to the application but not show them.

Spanish (Latin America) was added, as taken from Transifex.
By default it will not be visible, so the application will behave exactly
like it does today.

related to PR #44