Remove old references to captcha
simple_captcha was already removed before from version 0.8 (see a890d345)
and was replaced by recaptcha. However, this was never done againg in the
new version.

Also, there are references to captcha in the code but it was not being used.
So I'm just removing all the code until we implement it again, properly.

A database migrate from Mconf-Web 0.8 was becoming different from a database
created directly on Mconf-Web 2.0. Removing the captcha tables and columns
solves this issue.

refs #727, refs #369