Avoid inexistent profile right after a user is created
A profile should be created right after the user is created, but the way it
was creating the profile was leaving room for a profile of a user to be
accessed before the user was created but after the profile was created.
Was seen on the fork mconf-rnp/mconf-web.

The way it was done now, the profile is always available when accessed.

See also: d292a33ca21873ff89ef1bd5757fd0ab336bafaa