Strip domain from usernames that are emails for new LDAP users
When a user is being created after a sign in via LDAP, it will now auto
remove the "@domain" part of the email in case the attribute in the username
field is an email.
It's useful for cases where we want to use the same username a user has for
his email on Mconf-Web, or when there's no username attribute on LDAP, only
the an email attribute.