Add matcher to check that a user can do everything over a model
There was already a matcher "be_able_to_do_anything_to", but it worked only
with "should_not". Added a new matcher "be_able_to_do_everything_to" that
works only for "should" and is more complete than using "be_able_to(:manage)".

Making a matcher that works for both cases it's not really possible because
the concepts are different (one is not simply a negation of the other).