In the test for UserNotificationWorker the activity it was not created because the logic by refs #1737 changed
In this test it was called the method shibboleth.create_(user) that before create a notification (a activity).
But the notification is needed to be created after both token and user are saved (by refs #1737), and how the token is only saved in the shibboleth controller,
the call for create_notification is not anymore in tre create_user method.
That's why the test for UserNotificationsWorker didn't pass, because before the failed test it called the create_user and not a crete_notification
after the token is saved. And now this call is added.