Small improvements on cropping small images, after PR #92
Move decision of whether an image is too small to the models.

Also redirect to profiles#edit whenever the profile is updated.
Was redirecting to profiles#show when the image was cropped, but not when
the image was too small or when the image was uploaded but not cropped.
Now it always renders profiles#edit after the update.

refs #1465