Review/Organize permissions for unapproved resources and add tests
Moved all restrictions over unapproved resources to a new method in the
abilities, the same way it is done for disabled resources.
Also reviewed some resources that were not being restricted when the owner
was disabled (rooms and recordings). Added tests for the new restrictions
added and for other things that were already implemented.

refs #1650