Stop using cancancan aliases and review permissions for :index
Removed the aliases that cancancan adds by default. Using :read, specially,
makes it a bit confusing in cases when the ability is defined with a block.
The block is evaluated for :show, but not for :index.
Now there are no more aliases, actions have to be defined explicitly.
Also, abilities for :index are now always defined without a block, so we
don't mistakenly think that the block is being evaluated.