Improve how the test LDAP server and the Site are configured
Some tests for LDAP were not passing because of inconsistencies between
the configuration in the site and the configurations in the fake LDAP server.
Made a few changes on Mconf::LdapServer to make the configurations more
clear and in separate methods that can be used outside of this class.
Also improved a few tests to use these new methods.

refs #1368