Add new_account to LdapToken to track users created by LDAP
The same way we have ShibToken#new_account to track users that were created
by Shibboleth.
It makes so that only users that were created by LDAP cannot sign in
locally. So a user can have an LdapToken but still be able to sign in
locally, in case the token was created for an account that already existed.

refs #1155