Move methods to generate dial numbers to bigbluebutton_rails
The methods to generate dial numbers are now in the gem. There's an action
in BigbluebuttonRails::RoomController that generates a unique random dial
number and saves in the room.
Removed from Mconf-Web some code that wasn't needed anymore and adapted
some other code.

Now dial numbers are not generated automatically when rooms are created.
They will only be generated when an admin explicitly clicks on the
"generate dial number" button in the management list of rooms.

refs #1583
refs #1635