Add flag to disable spaces (first step) and new way to disable mods
Added the flag and made some basic configurations to disable resources
related to spaces if the flag is disabled. Not everything yet, there's
still a lot to do.

Also improved the way "modules" are disabled to make it simpler and cleaner
in the application. Now disabling a mod means:
* Routes are disabled.
* There's a default scope in the module that returns nothing, making sure
  resources are never found even if they exist in the db.
* Views hide what's needed.

There's no need change anything else. So code in the controllers is always
the same, no matter if the module is enabled or not, and it will work.

refs #1908