Move some code from MemberAbility to BaseAbility after PR #176
There's already a place in BaseAbility to remove access from things related
to disabled and unapproved resources.
The code got bigger, but mostly because of a test that tries to set an
invalid class name as owner_type in the event. But also because it now
restrict :new and :create over events of unapproved and disabled users.

refs #1928