Add migration to set default values to flags removed from rooms
We want to reset the flags now that users will not be able to access them
anymore.
The way flags were set before was not ideal. The user could save the room,
not changing any of the flags, and it would cause the flags to change from
nil to false. That had an implication in the conference that users were
not aware of. So resetting the flags will make things more consistent.

Also changed the way these flags are edited in the management page from
checkboxes to selects. Selects allow 3 values, that include the nil option
and fix the issue with inadvertently setting the flags when saving a room.

refs #1811