Dont access .users for each space in space index
In spaces#index, to define if a user is a member of a space we were testing for inclusion of user_id in space.user_ids. This is a lot slower because for each space we do a join to query for the users. A better way is getting a list of user spaces and testing each space_id for inclusion.