Fix setting user as signed in via Shibboleth
Was setting the user as signed in via shib after any call to `load_data`.
But that alone doesn't mean the user has signed in, since the data can
be empty or invalid.
Now there's an specific method to be called when a user is confirmed to be
signed in via shib. It will then store this information in the session.

refs #1895