Fix how recent activities are created for meetings
One error is that it was trying to get information from the controller,
but since meetings are usually created from models/workers, the controller
would be nil. This fixes issue #861.

There were tests to make sure the RA was correct, but they would always
pass if running after controller tests, and that's why the error was not
detected before.

Updating the gem public_activity fixed the conflicts in the tests.
With the update, also disabled recent activity creation by default for all
tests (expect on features and workers), which improves performance a bit.
Now tests that require activities have to require them explicitly.