Improve redirects and links to unpublish/destroy recordings
Removed the hack from the recordings controller and replaced by parameters
that are used by BigbluebuttonRails to redirect the user back.

Added confirmations on links to destroy/unpublish recordings.

Also changed the "destroy" link shown to admins in lists of recordings
to a "manage" link, that redirects to a management page that has links to
all actions possible (including the "destroy").