Don't allow access to /login when local auth is disabled
It would allow users to access the route but not allow the sign in unless
the user was an admin. Now it blocks the route, redirecting to root_path.
There's also a new route "/admin" specific for admins to sign in when local
auth is disabled.

This change was made because the application redirects to /login in several
places even if local auth is disabled, and it doesn't make sense in some
cases. It's easier to redirect /login to another place than to find all
redirects to /login and check beforehand that local auth is enabled.