Improve setting rooms to public/private, use just one key (attendee)
Now users cannot set one attendee and one moderator key in a room, they can
only set a single key, for attendees.
The form to change the visibility of rooms was improved to show/hide the
input for the key according to the visibility chosen.
It's easier for the user to understand what's happening now.