Encapsulate guest login into a module and add several helper methods
The logic to sign in using certificate without creating an account is now
encapsulated in a module that controls "guest users", users that are
authenticated somehow but do not have an account.

There are options to sign in and sign out guest users similarly to how
devise does it.

Review for PR #939.