Fix small font sizes in emails
Some email readers would show the emails with really small letters. In the
end it was just because we were setting the font-size to something like
"0.8125rem" and this would not play well with some email clients.

Removing the font-size lets the email client decide the appropriate size
and now emails look better in all clients.

Fixes #814