Auto open modal windows after trying to access them not via xhr
If a route that renders a modal is accessed with a normal request (not xhr),
it redirects the user back to the previous page. Now it will also set a
parameter in the URL and automatically open the modal window after
redirecting.

Also improved the view /join_requests/new to show better messages and links
when the user already has a pending join request.