Remove join_requests/show, not necessary anymore
Now join requests are accepted/declined in the user's home page and in the
space's administration pages only, there no show view for a join request.
The view was redundant, since users can respond to join requests in other
places, and had some issues (mostly with redirects). We opted for removing
it and to plan on improving the way users see join requests in the future
(e.g. using internal notifications).