Implement record permissions on join instead of create
Now the flag that controls whether a user can record or not is in the join
API call and not in the create anymore. This makes the permission control
a lot more flexible.

This feature is currently only on Mconf-Live, so it is disabled in
Mconf-Web. It can be enabled by setting to true the flag:
    "Rails.application.config.per_user_record_permissions"