Use 'rerequest' on facebook's auth to deal with users not sharing email
If the user tried to authenticate via Facebook and didn't share the email,
the login would fail and the user would have to sign in to Facebook and
remove the Mconf app to be able to try to login on Mconf again.

Now it will request users to share their email when they click in the
sign in button if they didn't share yet.